added auth checks to createPost + deletePost controllers - valid user only

+++ controllers/feed.js
@@ -71,7 +71,8 @@ exports.createPost = (req, res, next) => {
             res.status(201).json({
                 message: 'Post created successfully!',
                 post,
-                creator: { _id: creator._id, name: creator.name }
+                creator: { _id: creator._id, name: creator.name },
+                
             });
         })
         .catch(err => {
@@ -127,6 +128,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -158,7 +164,11 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
-            // Check for logged in user
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
